# Conflicts:
#	.gradle/checksums/checksums.lock
#	.gradle/checksums/md5-checksums.bin
#	.gradle/checksums/sha1-checksums.bin
#	.idea/workspace.xml

+++ src/main/java/com/project/devidea/modules/content/mentoring/MenteeService.java
@@ -1,36 +1,80 @@
 package com.project.devidea.modules.content.mentoring;
 
 import com.project.devidea.modules.account.Account;
+import com.project.devidea.modules.content.mentoring.exception.AlreadyExistException;
+import com.project.devidea.modules.content.mentoring.exception.NotFoundException;
+import com.project.devidea.modules.content.mentoring.form.CreateMenteeRequest;
 import com.project.devidea.modules.content.mentoring.form.UpdateMenteeRequest;
+import com.project.devidea.modules.tagzone.tag.Tag;
+import com.project.devidea.modules.tagzone.tag.TagRepository;
+import com.project.devidea.modules.tagzone.zone.Zone;
+import com.project.devidea.modules.tagzone.zone.ZoneRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.util.Set;
+
 @Service
 @Transactional
-@RequiredArgsConstructor
-public class MenteeService {
+public class MenteeService extends AbstractService {
 
     private final MenteeRepository menteeRepository;
 
-    public Long createMentee(Mentee mentee, Account account) {
-        mentee.setAccount(account);
-        mentee.publish();
-        menteeRepository.save(mentee);
-        return mentee.getId();
+    protected MenteeService(TagRepository tagRepository, ZoneRepository zoneRepository, MenteeRepository menteeRepository) {
+        super(tagRepository, zoneRepository);
+        this.menteeRepository = menteeRepository;
     }
 
-    public void updateMentee(UpdateMenteeRequest request, Mentee mentee) {
-        // modelMapper.map(request, mentee);
+    public Long createMentee(Account account, CreateMenteeRequest request) {
+
+        if(getMentee(account.getId()) != null) {
+            throw new AlreadyExistException("이미 등록된 멘티입니다.");
+        }
+
+        Set<Zone> zones = getZones(request.getZones());
+        Set<Tag> tags = getTags(request.getTags());
+
+        Mentee mentee = Mentee.createMentee(account, request.getDescription(), zones, tags, request.isFree());
+        return menteeRepository.save(mentee).getId();
+    }
+
+    private Mentee getMentee(Long accountId) {
+        return menteeRepository.findByAccountId(accountId);
+    }
+
+    public void deleteMentee(Account account) {
+
+        Mentee mentee = getMentee(account.getId());
+        if(mentee == null) {
+            throw new NotFoundException("존재하지 않는 멘티입니다.");
+        }
+
+        mentee.getTags().clear();
+        mentee.getZones().clear();
+        mentee.setAccount(null);
+        menteeRepository.delete(mentee);
+    }
+
+    public void updateMentee(Account account, UpdateMenteeRequest request) {
+
+        Mentee mentee = getMentee(account.getId());
+        if(mentee == null) {
+            throw new NotFoundException("존재하지 않는 멘티입니다.");
+        }
+
+        Set<Zone> zones = getZones(request.getZones());
+        Set<Tag> tags = getTags(request.getTags());
+
         mentee.setDescription(request.getDescription());
-        mentee.setZones(request.getZones());
-        mentee.setTags(request.getTags());
+
+        mentee.getZones().clear();
+        mentee.setZones(zones);
+        mentee.getTags().clear();
+        mentee.setTags(tags);
+
         mentee.setOpen(request.isOpen());
         mentee.setFree(request.isFree());
     }
 
-    public void deleteMentee(Mentee mentee) {
-        mentee.setAccount(null);
-        menteeRepository.delete(mentee);
-    }
 }
